fix(conform-zod): use parseFloat not parseInt

+++ packages/conform-zod/coercion.ts
@@ -121,7 +121,7 @@ export function enableTypeCoercion<Type extends ZodTypeAny>(
 			.pipe(type);
 	} else if (type instanceof ZodNumber) {
 		schema = any()
-			.transform((value) => coerceString(value, parseInt))
+			.transform((value) => coerceString(value, parseFloat))
 			.pipe(type);
 	} else if (type instanceof ZodBoolean) {
 		schema = any()
